Refactor and introduce spi.ac package
