+++ digdag-server/src/main/java/io/digdag/server/ac/DefaultAccessController.java
@@ -1,6 +1,6 @@
 package io.digdag.server.ac;
 
-import io.digdag.spi.AccessController;
+import io.digdag.spi.ac.AccessController;
 
 public class DefaultAccessController
         implements AccessController
