+++ digdag-spi/src/main/java/io/digdag/spi/ac/AccessControlException.java
@@ -1,4 +1,4 @@
-package io.digdag.spi;
+package io.digdag.spi.ac;
 
 /**
  * An exception thrown when a required action to a target (project, revision, session, etc.) is not be permitted.
